requirements.txt
wheel
aiortc
aiohttp==3.6.3
av
numpy
uvloop
EasyProcess
Pillow
mss
python-uinput
python-xlib
uvloop
upnpy
opencv-python
